backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
snowflake-connector-python>=3.6.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0
